+++ esp-hal-procmacros/src/rtos_main.rs
@@ -191,7 +191,7 @@ pub fn main_fn() -> TokenStream2 {
             let mut executor = ::esp_rtos::embassy::Executor::new();
             let executor = unsafe { __make_static(&mut executor) };
             executor.run(|spawner| {
-                spawner.must_spawn(__embassy_main(spawner));
+                spawner.spawn(__embassy_main(spawner).unwrap());
             })
         }
     }
@@ -233,7 +233,7 @@ mod tests {
                         let mut executor = ::esp_rtos::embassy::Executor::new();
                         let executor = unsafe { __make_static (& mut executor) };
                         executor . run (| spawner | {
-                            spawner.must_spawn(__embassy_main (spawner));
+                            spawner.spawn(__embassy_main (spawner).unwrap());
                         })
                     }
                 }
@@ -389,7 +389,7 @@ mod tests {
                         let mut executor = ::esp_rtos::embassy::Executor::new();
                         let executor = unsafe { __make_static (& mut executor) };
                         executor.run (| spawner | {
-                            spawner.must_spawn(__embassy_main (spawner));
+                            spawner.spawn(__embassy_main (spawner).unwrap());
                         })
                     }
                 }
@@ -430,7 +430,7 @@ mod tests {
                         let mut executor = ::esp_rtos::embassy::Executor::new();
                         let executor = unsafe { __make_static (& mut executor) };
                         executor.run (| spawner | {
-                            spawner.must_spawn(__embassy_main (spawner));
+                            spawner.spawn(__embassy_main (spawner).unwrap());
                         })
                     }
                 }
@@ -486,7 +486,7 @@ mod tests {
                         let mut executor = ::esp_rtos::embassy::Executor::new();
                         let executor = unsafe { __make_static (& mut executor) };
                         executor.run (| spawner | {
-                            spawner.must_spawn(__embassy_main (spawner));
+                            spawner.spawn(__embassy_main (spawner).unwrap());
                         })
                     }
                 }